+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/result/SearchResultFactory.kt
@@ -99,6 +99,7 @@ internal class SearchResultFactory(private val dataProviderResolver: DataProvide
             OriginalResultType.POI,
             OriginalResultType.STREET,
             OriginalResultType.POSTCODE,
+            OriginalResultType.BLOCK,
             OriginalResultType.QUERY -> {
                 when (apiType) {
                     ApiType.GEOCODING -> {